requirements.txt
# Core data processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0       # Required by BigQuery client for DataFrame uploads

# Google Cloud
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0  # Faster reads via Storage API
google-auth==2.30.0

# HTTP requests (API simulation)
requests==2.31.0

# Airflow (orchestration)
apache-airflow==2.9.1

# Claude API (conversational agent)
anthropic==0.34.0

# Logging & utilities
python-dotenv==1.0.1
colorlog==6.8.2

# Testing
pytest==8.2.2
pytest-cov==5.0.0

# Data validation
great-expectations==0.18.15
